Rename SolverTalbe to SolverTableEntry
Added Skipped and Retries so that we can keep track of multiple attempts to do the same thing.

Made logger a global since it should be used everywhere and shouldn't need to be passed to everyting

Added outputTable since that's the ultimate goal here.

Add a check at the beginning and every few seconds to see if the solution has been found and we should exit